Implements ui behavior for small screens
