+++ src/pages/monitoring/outlets/initiativesMap/utils/mapPositioning.ts
@@ -1,8 +1,35 @@
 import L from "leaflet";
 
-import { COUNTRY_BOUNDS } from "@config/monitoring";
+import {
+  COUNTRY_BOUNDS,
+  INITIATIVES_MAP_PADDING_SM,
+  INITIATIVES_MAP_PADDING_LG,
+} from "@config/monitoring";
 
-const padding = { north: 0, south: 0, east: 100, west: 400 };
+const padding = {
+  get south() {
+    return typeof window !== "undefined" && window.innerWidth < 1024
+      ? INITIATIVES_MAP_PADDING_SM.south
+      : INITIATIVES_MAP_PADDING_LG.south;
+  },
+  get north() {
+    return typeof window !== "undefined" && window.innerWidth < 1024
+      ? INITIATIVES_MAP_PADDING_SM.north
+      : INITIATIVES_MAP_PADDING_LG.north;
+  },
+  get east() {
+    return typeof window !== "undefined" && window.innerWidth < 1024
+      ? INITIATIVES_MAP_PADDING_SM.east
+      : INITIATIVES_MAP_PADDING_LG.east;
+  },
+  get west() {
+    return typeof window !== "undefined" && window.innerWidth < 1024
+      ? INITIATIVES_MAP_PADDING_SM.west
+      : INITIATIVES_MAP_PADDING_LG.west;
+  },
+};
+
+// const padding = { south: 100, north: 0, east: 20, west: 20 };
 
 /**
  * Calculates the optimal center and zoom level to fit specific bounds within
@@ -22,7 +49,7 @@ export const getTargetBounds = (
   const boundsObject = L.latLngBounds(baseBounds);
   const paddingPoint = L.point(
     padding.east + padding.west,
-    padding.north + padding.south,
+    padding.south + padding.north,
   );
 
   const targetZoom = targetMap.getBoundsZoom(boundsObject, false, paddingPoint);
@@ -31,7 +58,7 @@ export const getTargetBounds = (
   const targetPoint = targetMap.project(baseCenter, targetZoom);
 
   const offsetX = (padding.east - padding.west) / 2;
-  const offsetY = (padding.north - padding.south) / 2;
+  const offsetY = (padding.south - padding.north) / 2;
   const offsetPoint = targetPoint.add([offsetX, offsetY]);
 
   const zoomCenter = targetMap.unproject(offsetPoint, targetZoom);
@@ -51,7 +78,7 @@ export const getTargetLatLng = (targetMap: L.Map): L.LatLng => {
   const mapSize = targetMap.getSize();
 
   const vcX = padding.west + (mapSize.x - padding.east - padding.west) / 2;
-  const vcY = padding.north + (mapSize.y - padding.north - padding.south) / 2;
+  const vcY = padding.south + (mapSize.y - padding.south - padding.north) / 2;
   const visualCenterPoint = L.point(vcX, vcY);
   const targetLatLng = targetMap.containerPointToLatLng(visualCenterPoint);
